requirements.txt
Django>=4.2.0,<4.3
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-ratelimit>=4.1.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
numpy>=1.24.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
dj-database-url>=2.1.0
whitenoise>=6.6.0
django-filter>=24.0
celery[redis]>=5.2.0
